resolved auto generated swagger ui rendering bug

+++ package/config/setting.js
@@ -1,8 +1,8 @@
 const swaggerUi = require('swagger-ui-express')
 const fs = require("fs")
-const {resolve } = require("path");
+const {dirname, join } = require("path");
 
-// const getRootPath = require("../util/getRootPath")
+const getRootPath = require("../util/getRootPath")
 const getMetadata = require("./getMetadata")
 
 
@@ -21,7 +21,7 @@ const config = (...options)=>{
     } = options[0]
     
     let docs = docsPath ? docsPath : 'api-docs'
-    // let autoGenPath =  join(getRootPath(dirname),"autoGens")
+    let autoGenPath =  join(getRootPath(dirname),"autoGens")
 
     const templete = getMetadata({
         title,
@@ -34,27 +34,51 @@ const config = (...options)=>{
     })
  
    
-   let path = resolve('../package/constants/swaggerTemplete.json')
-   
-    fs.writeFileSync(
-        path, 
+
+   if(!fs.existsSync(autoGenPath)){
+       fs.mkdirSync(autoGenPath, {recursive:true})
+       
+        fs.writeFileSync(
+        join(autoGenPath, "swaggerDocument.json"), 
         JSON.stringify(templete),
         (err)=>{
             if(err) console.log(err)
-        } 
-    )
-
-    if (environment=== "production"){
-        let file = require('../schemas/swaggerDocument.json')
-        app.use(`/${docs}`, swaggerUi.serve, swaggerUi.setup(
-            file
-        ))
-    }
-    else{
-        fs.readFile(path, "utf8", (err, file)=>{
+            } 
+        )
+        fs.writeFileSync(
+        join(autoGenPath, "pathSchema.json"), 
+        JSON.stringify({}),
+        (err)=>{
+            if(err) console.log(err)
+            } 
+        )
+        fs.writeFileSync(
+        join(autoGenPath, "jsonSchema.json"), 
+        JSON.stringify({}),
+        (err)=>{
+            if(err) console.log(err)
+            } 
+        ) 
+        fs.readFile(join(autoGenPath, "swaggerDocument.json"), "utf8", (err, file)=>{
             if(err)console.log(err)
             app.use(`/${docs}`, swaggerUi.serve, swaggerUi.setup(JSON.parse(file)))
         }) 
+
+       
+   }
+    else{
+        if (environment=== "production"){
+            let file = require(join(autoGenPath, "swaggerDocument.json"))
+            app.use(`/${docs}`, swaggerUi.serve, swaggerUi.setup(
+                file
+            ))
+        }
+        else{
+            fs.readFile(join(autoGenPath, "swaggerDocument.json"), "utf8", (err, file)=>{
+                if(err)console.log(err)
+                app.use(`/${docs}`, swaggerUi.serve, swaggerUi.setup(JSON.parse(file)))
+            }) 
+        }
     }
 
 }
